+++ Vector.java
@@ -1,9 +1,9 @@
 
 public class Vector {
 	
-	public final float x,y,z;
+	public final double x,y,z;
 		
-	public Vector(float x, float y, float z)
+	public Vector(double x, double y, double z)
 	{
 		this.x = x;
 		this.y = y;
@@ -29,7 +29,7 @@ public class Vector {
 		return new Vector(v1.y*v2.z - v1.z*v2.y, v1.z*v2.x-v1.x*v2.z,v1.x*v2.y-v1.y*v2.x);
 	}
 	
-	public static float dotProduct(Vector first, Vector second)
+	public static double dotProduct(Vector first, Vector second)
 	{
 		return (first.x*second.x + first.y*second.y + first.z*second.z);
 	}
@@ -48,27 +48,27 @@ public class Vector {
 		return new Vector(this.x + second.x, this.y + second.y, this.z + second.z);
 	}
 	
-	public Vector scalarProduct(float scalar)
+	public Vector scalarProduct(double scalar)
 	{
 		return new Vector(scalar*this.x, scalar*this.y, scalar*this.z);
 	}
 	
 	public Vector getProjection(Vector vector)
 	{
-		float dotproduct = Vector.dotProduct(this,vector);
+		double dotproduct = Vector.dotProduct(this,vector);
 		dotproduct /= vector.getLengthSquare();
 		Vector normalized = vector.scalarProduct(dotproduct);
 		return normalized;
 	}
 	
-	public float getLengthSquare()
+	public double getLengthSquare()
 	{
 		return (this.x*this.x + this.y*this.y + this.z*this.z);
 	}
 	
-	public float getLength()
+	public double getLength()
 	{
-		return (float)Math.sqrt(this.getLengthSquare());
+		return Math.sqrt(this.getLengthSquare());
 	}
 	
 	public String toString()
